chore: remove facebook graceful shutdown package.
